feat: add ProspectBottom component
