+++ packages/apps/crm/src/components/organisms/index.js
@@ -6,3 +6,4 @@ export {default as LeadBody} from './LeadBody/LeadBody';
 export {default as LeadBottom} from './LeadBottom/LeadBottom';
 export {default as ProspectHeader} from './ProspectHeader/ProspectHeader';
 export {default as ProspectBody} from './ProspectBody/ProspectBody';
+export {default as ProspectBottom} from './ProspectBottom/ProspectBottom';
