+++ packages/apps/crm/src/screens/prospect/ProspectDetailsScreen.js
@@ -1,12 +1,12 @@
 import React, {useEffect} from 'react';
-import {ScrollView, StyleSheet, View} from 'react-native';
-import {Screen, HeaderContainer, CircleButton} from '@axelor/aos-mobile-ui';
+import {ScrollView} from 'react-native';
+import {Screen, HeaderContainer} from '@axelor/aos-mobile-ui';
 import {
   useSelector,
   HeaderOptionsMenu,
   useDispatch,
 } from '@axelor/aos-mobile-core';
-import {ProspectBody, ProspectHeader} from '../../components';
+import {ProspectBody, ProspectBottom, ProspectHeader} from '../../components';
 import {searchContactById} from '../../features/contactSlice';
 import {fetchPartnerEventById} from '../../features/eventSlice';
 import {fetchProspectById} from '../../features/prospectSlice';
@@ -53,58 +53,9 @@ const ProspectDetailsScreen = ({navigation, route}) => {
       <ScrollView>
         <ProspectBody navigation={navigation} />
       </ScrollView>
-      <View style={styles.bottomContainer}>
-        <CircleButton
-          iconName="pen"
-          onPress={() =>
-            navigation.navigate('ProspectFormScreen', {
-              idProspect: idProspect,
-            })
-          }
-        />
-      </View>
+      <ProspectBottom idProspect={idProspect} navigation={navigation} />
     </Screen>
   );
 };
-
-const styles = StyleSheet.create({
-  item: {
-    marginHorizontal: 12,
-    marginVertical: 4,
-  },
-  headerContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-  headerContainerChildren: {
-    flexDirection: 'row',
-    marginLeft: '5%',
-    alignItems: 'center',
-  },
-  headerInfo: {
-    flexDirection: 'column',
-    marginLeft: '7%',
-  },
-  headerBadge: {
-    flexDirection: 'column',
-  },
-  container: {
-    alignItems: 'center',
-  },
-  textTitle: {
-    fontWeight: 'bold',
-  },
-  leadScoring: {
-    marginTop: '10%',
-  },
-  bottomContainer: {
-    width: '90%',
-    flexDirection: 'row-reverse',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginVertical: 2,
-    marginBottom: 25,
-  },
-});
 
 export default ProspectDetailsScreen;
